# api-gateway/docs/swagger.yaml
basePath: /
definitions:
  genproto.LogCreateRequest:
    properties:
      level:
        type: string
      message:
        type: string
      service_name:
        type: string
    type: object
  genproto.LogGetAll:
    properties:
      count:
        type: integer
      log:
        items:
          $ref: '#/definitions/genproto.LogGetResponse'
        type: array
    type: object
  genproto.LogGetResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      id:
        type: integer
      level:
        type: string
      message:
        type: string
      service_name:
        type: string
    type: object
  genproto.LogUpdateRequest:
    properties:
      id:
        type: integer
      level:
        type: string
      message:
        type: string
      service_name:
        type: string
    type: object
...
  genproto.PostCreateRequest:
    properties:
      content:
        type: string
      title:
        type: string
      user_id:
        type: integer
    type: object
  genproto.PostGet:
    properties:
      content:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      id:
        type: integer
      title:
        type: string
      user:
        $ref: '#/definitions/genproto.User'
      user_id:
        type: integer
    type: object
  genproto.PostGetAll:
    properties:
      count:
        type: integer
      post:
        items:
          $ref: '#/definitions/genproto.PostGet'
        type: array
    type: object
  genproto.PostGetResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      id:
        type: integer
      title:
        type: string
      user_id:
        type: integer
    type: object
  genproto.PostUpdateRequest:
    properties:
      content:
        type: string
      id:
        type: integer
      title:
        type: string
      user_id:
        type: integer
    type: object
  genproto.User:
    properties:
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phone_number:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: API for Instant Delivery resources
  title: NDC Post Project API Documentation
  version: "1.0"
paths:
...
  /api/v1/logs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a log by its ID
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Log deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Log not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Log
      tags:
      - Log
    get:
      consumes:
      - application/json
      description: Retrieve a log by its ID
      parameters:
      - description: Log ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Log details
          schema:
            $ref: '#/definitions/genproto.LogGetResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Log not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Log
      tags:
      - Log
  /api/v1/logs/create:
    post:
      consumes:
      - application/json
      description: Create a new log
      parameters:
      - description: Log data
        in: body
        name: log
        required: true
        schema:
          $ref: '#/definitions/genproto.LogCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Log created successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Log
      tags:
      - Log
  /api/v1/logs/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of logs with pagination
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: Level
        in: query
        name: level
        type: string
      - description: ServiceName
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of logs
          schema:
            $ref: '#/definitions/genproto.LogGetAll'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Logs
      tags:
      - Log
  /api/v1/logs/update:
    patch:
      consumes:
      - application/json
      description: Update an existing log
      parameters:
      - description: Log data
        in: body
        name: log
        required: true
        schema:
          $ref: '#/definitions/genproto.LogUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Log updated successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Log
      tags:
      - Log
  /api/v1/posts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Post not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete Post
      tags:
      - Post
    get:
      consumes:
      - application/json
      description: Retrieve a post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Post details
          schema:
            $ref: '#/definitions/genproto.PostGetResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: Post not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Post
      tags:
      - Post
  /api/v1/posts/create:
    post:
      consumes:
      - application/json
      description: Create a new post
      parameters:
      - description: Post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/genproto.PostCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Post created successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create Post
      tags:
      - Post
  /api/v1/posts/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of posts with pagination
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      - description: UserID
        in: query
        name: user_id
        type: integer
      - description: Content
        in: query
        name: content
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of posts
          schema:
            $ref: '#/definitions/genproto.PostGetAll'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get Posts
      tags:
      - Post
  /api/v1/posts/update:
    patch:
      consumes:
      - application/json
      description: Update an existing post
      parameters:
      - description: Post data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/genproto.PostUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Post updated successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update Post
      tags:
      - Post
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"